requirements.txt
pdf2image
Pillow==9.5.0
PySimpleGUI
python-poppler